Lp Pool Migration

+++ crates/lp-pool/src/block_header.rs
@@ -114,7 +114,7 @@ impl BlockHeader {
         let metadata_addr = data_addr
             .checked_sub(metadata_size)
             .expect("data pointer lacks metadata region");
-        let stored_ptr = core::ptr::read_unaligned(metadata_addr as *const usize) as usize;
+        let stored_ptr = core::ptr::read_unaligned(metadata_addr as *const usize);
         let block_addr = stored_ptr;
 
         debug_assert!(
